Remove the keyboard benchmark.
Too tied to the Racket infrastrcture, which makes it prone to breaking.
This would also make it hard to statically link, which is where these
benchmarks should be headed (dynamic linking via shadowing has been a
constant source of breakage and frustration).

Furthermore, this benchmark is annoying to run: it requires an extra
package to be installed (which I haven't found a good way to do
programmatically and without affecting the rest of the system) and takes
over the X session while it runs.

This would be a good benchmark, though, so we should revisit it
eventually.